Fix test. Move get_base_name
